% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Auto_EDA}
\alias{Auto_EDA}
\title{Perform Exploratory Data Analysis}
\usage{
Auto_EDA(dataset)
}
\arguments{
\item{dataset}{The pre-processed dataset to be analysed}
}
\description{
Generates a variety of plots and reports summarising the variable characteristics with simple user selections
}
