% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_Proc.R
\name{Pre_Proc}
\alias{Pre_Proc}
\title{Pre-process dataset}
\usage{
Pre_Proc(dataset)
}
\arguments{
\item{dataset}{The dataset to be analysed. The dataset should be structured with column names preferably present}
}
\value{
The pre-processed dataset
}
\description{
GUI assisted data transformation and manipulation, outlier detection and missing value treatment
}
